Move eacp onto EA::Vector/Array/Span and int-sized interfaces
Replace every std::array, std::vector and std::span in Lib, Apps and Tests
with the EA containers re-exported by Core/Utils/Containers.h, and use int
instead of size_t across eacp-facing interfaces. size_t stays only where the
language, the STL or an OS API forces it: memcpy lengths, hash functors,
operator new, C-array extents, std::string indexing, and the mmap, Core
Graphics, Metal, CoreText, libcurl, Winsock and D3D12 call sites. The D3D12
allocator in GPU/Windows/D3D12Context.h keeps std::size_t internally.

Files that can outgrow an int use the size_t surface the library now offers:
MemoryMappedFile reports a size_t size and hands out a Span over the whole
mapping, and the MP4 BoxReader walks boxes with a size_t cursor, so files
over 2 GB map and demux. A test maps a sparse file past the int boundary.

Behaviour changes made on the way: the HTTP request parser clamps a
Content-Length beyond 2^31-1 instead of believing it, GPU::Buffer::read and
update reject negative sizes, WebSocket message limits are compared in
64-bit, and Text::registerMemoryFont rejects a null or non-positive size.

Requires ea_data_structures 2cdfc28 (getSize() on the containers, Span over
std::span). Windows and Linux sources were converted without a compiler for
them; CI is their first check.

+++ Lib/eacp/WebView/WebView/WebView.mm
@@ -654,8 +654,7 @@ void schemeHandlerPumpTask(id self,
 
       auto got = context->resource.read(
           context->offset,
-          eacp::Graphics::ByteSpan {context->buffer.data(),
-                                    static_cast<std::size_t>(want)});
+          eacp::Graphics::ByteSpan {context->buffer.data(), (int) want});
 
       dispatch_async(dispatch_get_main_queue(), ^{
         if (! live->load())
